Refactor Person and add Member class
Updated Library.csproj to include Member.cs.
Changed BirthDate in Person.cs to a concrete property with validation.
Removed BirthDate from User.cs.
Added Member class in Member.cs under Library.Model.Concrete.
